Revert "Migrated functional tests from a separate git project"
