--- api/src/functionalTest/java/uk/gov/hmcts/payment/functional/IntegrationTestBase.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.payment.functional;
-
-import org.junit.runner.RunWith;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@ContextConfiguration(classes = TestContextConfiguration.class)
-public class IntegrationTestBase {
-}
