--- api/src/functionalTest/java/uk/gov/hmcts/payment/functional/TestContextConfiguration.java
@@ -1,11 +1,0 @@
-package uk.gov.hmcts.payment.functional;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-@ComponentScan("uk.gov.hmcts.payment.functional")
-@PropertySource("classpath:application-functional-tests.properties")
-public class TestContextConfiguration {
-}
