+++ api/src/functionalTest/java/uk/gove/hmcts/fees/register/functional/FunctionalTest.java
@@ -1,38 +1,34 @@
-package uk.gov.hmcts.fees.register.smoke;
+package uk.gove.hmcts.fees.register.functional;
 
 import io.restassured.RestAssured;
-import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.junit4.SpringRunner;
-import static org.hamcrest.Matchers.equalTo;
 
 import static io.restassured.RestAssured.given;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 
 @RunWith(SpringRunner.class)
-@Slf4j
-public class SmokeTest {
+public class FunctionalTest {
+
     @Value("${TEST_URL:http://localhost:8080}")
     private String testUrl;
 
     @Before
     public void setup() {
         RestAssured.baseURI = testUrl;
-        log.info("Payments-Api base url is :{}", testUrl);
     }
 
     @Test
-    public void healthCheck() {
+    public void shouldReturnChannels() {
         given()
             .relaxedHTTPSValidation()
             .header(CONTENT_TYPE, "application/json")
             .when()
-            .get("/health")
+            .get("/refdata/channels")
             .then()
-            .statusCode(200)
-            .body("status", equalTo("UP"));
+            .statusCode(200);
     }
 }
