--- api/src/functionalTest/resources/application-functional-tests.properties
@@ -1,3 +1,0 @@
-base-urls.payment=${TEST_URL:http://localhost:8080}
-base-urls.idam=https://idam-test.dev.ccidam.reform.hmcts.net
-base-urls.service-auth-provider=${S2S_URL}
