--- api/src/functionalTest/resources/gov-pay-stub/mappings/cancel-payment.json
@@ -1,12 +1,0 @@
-{
-  "request": {
-    "method": "POST",
-    "url": "/govPayId1/cancel"
-  },
-  "response": {
-    "status": 204,
-    "headers": {
-      "Content-Type": "application/json"
-    }
-  }
-}
