--- api/src/functionalTest/resources/gov-pay-stub/mappings/get-payment.json
@@ -1,61 +1,0 @@
-{
-  "request": {
-    "method": "GET",
-    "url": "/govPayId1"
-  },
-  "response": {
-    "status": 200,
-    "headers": {
-      "Content-Type": "application/json"
-    },
-    "jsonBody": {
-      "amount": 100,
-      "state": {
-        "status": "created",
-        "finished": false
-      },
-      "description": "Description",
-      "reference": "Reference",
-      "payment_id": "govPayId1",
-      "payment_provider": "sandbox",
-      "return_url": "https://return-url",
-      "created_date": "2017-03-16T15:32:27.668Z",
-      "refund_summary": {
-        "status": "pending",
-        "amount_available": 55000,
-        "amount_submitted": 0
-      },
-      "settlement_summary": {},
-      "_links": {
-        "self": {
-          "href": "http://wiremock-api:8080/govPayId1",
-          "method": "GET"
-        },
-        "next_url": {
-          "href": "https://www.payments.service.gov.uk/secure/next-url",
-          "method": "GET"
-        },
-        "next_url_post": {
-          "type": "application/x-www-form-urlencoded",
-          "params": {
-            "chargeTokenId": "b6a5fd11-2b74-411f-867b-f58789d0db8b"
-          },
-          "href": "https://www.payments.service.gov.uk/secure",
-          "method": "POST"
-        },
-        "events": {
-          "href": "http://wiremock-api:8080/govPayId1/events",
-          "method": "GET"
-        },
-        "refunds": {
-          "href": "http://wiremock-api:8080/govPayId1/refunds",
-          "method": "GET"
-        },
-        "cancel": {
-          "href": "http://wiremock-api:8080/govPayId1/cancel",
-          "method": "POST"
-        }
-      }
-    }
-  }
-}
