--- api/src/functionalTest/resources/gov-pay-stub/mappings/refund-payment-invalid-amount.json
@@ -1,21 +1,0 @@
-{
-  "request": {
-    "method": "POST",
-    "url": "/govPayId1/refunds",
-    "bodyPatterns": [
-      {
-        "equalToJson": "{ \"amount\": 5, \"refund_amount_available\": 1000 }"
-      }
-    ]
-  },
-  "response": {
-    "status": 412,
-    "headers": {
-      "Content-Type": "application/json"
-    },
-    "jsonBody": {
-      "code": "P0604",
-      "description": "Refund amount available mismatch"
-    }
-  }
-}
