--- api/src/functionalTest/resources/gov-pay-stub/mappings/refund-payment.json
@@ -1,17 +1,0 @@
-{
-  "request": {
-    "method": "POST",
-    "url": "/govPayId1/refunds",
-    "bodyPatterns": [
-      {
-        "equalToJson": "{ \"amount\": 5, \"refund_amount_available\": 100 }"
-      }
-    ]
-  },
-  "response": {
-    "status": 204,
-    "headers": {
-      "Content-Type": "application/json"
-    }
-  }
-}
