+++ api/src/smokeTest/java/uk/gove/hmcts/fees/register/smoke/SmokeTest.java
@@ -1,34 +1,38 @@
-package uk.gov.hmcts.fees.register.functional;
+package uk.gove.hmcts.fees.register.smoke;
 
 import io.restassured.RestAssured;
+import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.junit4.SpringRunner;
+import static org.hamcrest.Matchers.equalTo;
 
 import static io.restassured.RestAssured.given;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 
 @RunWith(SpringRunner.class)
-public class FunctionalTest {
-
+@Slf4j
+public class SmokeTest {
     @Value("${TEST_URL:http://localhost:8080}")
     private String testUrl;
 
     @Before
     public void setup() {
         RestAssured.baseURI = testUrl;
+        log.info("Payments-Api base url is :{}", testUrl);
     }
 
     @Test
-    public void shouldReturnChannels() {
+    public void healthCheck() {
         given()
             .relaxedHTTPSValidation()
             .header(CONTENT_TYPE, "application/json")
             .when()
-            .get("/refdata/channels")
+            .get("/health")
             .then()
-            .statusCode(200);
+            .statusCode(200)
+            .body("status", equalTo("UP"));
     }
 }
